requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.2.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
coverage==7.5.4
crispy-bootstrap5==2024.2
cssbeautifier==1.15.1
dill==0.3.8
Django==5.2.7
django-crispy-forms==2.1
django-djconfig==0.11.0
django-extensions==3.2.3
django-filter==25.2
django-formtools==2.5.1
django-haystack==3.3.0
django-infinite-scroll-pagination==1.1.0
django-js-asset==2.2.0
django-machina==1.3.1
django-mptt==0.16.0
django-qr-code==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djlint==1.35.4
EditorConfig==0.12.4
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==25.8.0
freezegun==1.5.5
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsbeautifier==1.15.1
json5==0.9.25
Markdown==3.10
markdown2==2.5.1
mccabe==0.7.0
mistune==0.8.4
olefile==0.46
packaging==24.0
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.2.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pylint-pytest==1.1.8
PyNaCl==1.5.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
segno==1.6.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.6
typing_extensions==4.11.0
tzdata==2024.2
whitenoise==6.6.0
Whoosh==2.7.4
